refactor: modules and setting up the application's working environment
