+++ platform/window/tgui/src/window.rs
@@ -1,0 +1,3 @@
+pub struct NativeWindowConfiguration {
+    pub full_screen: bool,
+}
