--- platform/workbench/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "workbench"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-configuration.workspace = true
-
-parking_lot.workspace = true
-lazy-regex.workspace = true
-anyhow.workspace = true
-arc-swap.workspace = true
-schemars.workspace = true
-tracing.workspace = true
-lazy_static.workspace = true
-hashbrown = { workspace = true, features = ["serde"] }
-
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
