--- platform/workbench/src/lib.rs
@@ -1,10 +1,0 @@
-pub mod service;
-pub mod tgui;
-
-#[macro_use]
-extern crate lazy_static;
-
-pub enum WorkbenchState {
-    Empty,
-    Workspace,
-}
