--- platform/workbench/src/service/configuration/common/configuration_editing.rs
@@ -1,4 +1,0 @@
-pub enum ConfigurationEditableTarget {
-    User,
-    Workspace,
-}
