--- platform/workbench/src/service/configuration/mod.rs
@@ -1,2 +1,0 @@
-pub mod common;
-pub mod tgui;
