--- platform/workbench/src/service/mod.rs
@@ -1,1 +1,0 @@
-pub mod configuration;
