--- platform/workbench/src/tgui/mod.rs
@@ -1,1 +1,0 @@
-pub mod workbench;
