--- platform/workbench/src/tgui/workbench.rs
@@ -1,1 +1,0 @@
-pub struct WorkbenchConfigurationService {}
