--- view/desktop/workbench/service/workspace_configuration/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod common;
-pub mod tgui;
