+++ view/shared/workbench/service/configuration/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "workspace_configuration"
+name = "workbench_service_configuration"
 version = "0.1.0"
 edition = "2021"
