+++ view/shared/workbench/service/configuration/src/common/workspace_configuration_model.rs
