+++ view/shared/workbench/service/configuration/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod common;
+pub mod tgui;
+pub mod workbench;
+
+#[macro_use]
+extern crate anyhow;
